pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "swarm_example"
version = "0.0.1"
authors = [{name = "Jan Willem Altink", email = "janwillem@janwillemaltink.eu"}]
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "python-dotenv",
    "openai",
    "swarm @ git+https://github.com/openai/swarm.git",
]

[project.scripts]
swarm_example = "swarm_example.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
